minio-manager to create users, policies and grant user the policies
Ensuring that minio client can login with user created with custom policy
It was hard to find, but with the current version of minio (may have been solved with more recent one if this issue is half related to https://github.com/minio/minio/issues/7696), some extra buckets needs to be accessible.
Indeed the client sends a request to a bucket `probe-bucket-sign-...` to validate that it can connect (on a `mc config host add ...`). Without, we get an `Access Denied` error. Great...

user username and password keys are optionals (with defaults)

rename a key in user object
userSecretName -> userCredentialsSecret

Using helm hooks to run jobs even on updates.